requirements.txt
Flask==2.1.0
google-cloud-datastore==2.9.0
protobuf==4.21.6